MegaBuy now closely monitors and reports on employee work habits.

+++ MegaBuy/Reports/WorkReportData.cs
@@ -4,13 +4,13 @@ namespace MegaBuy.Reports
 {
     public sealed class WorkReportData
     {
-        public float HoursWorkedActual { get; set; }
-        public float HoursWorkedExpected { get; set; }
+        public double HoursWorkedActual { get; set; }
+        public double HoursWorkedExpected { get; set; }
         public int CallsAnsweredActual { get; set; }
         public int CallsAnsweredExpected { get; set; }
-        public float AvgCallRatingActual { get; set; }
-        public float AvgCallRatingExpected { get; set; }
-
+        public double AvgCallRatingActual { get; set; }
+        public double AvgCallRatingExpected { get; set; }
+        
         public CallerFeedback BestFeedback { get; set; }
         public CallerFeedback WorstFeedback { get; set; }
     }
